+++ docs/changelog.rst
@@ -14,6 +14,8 @@ Not yet released
 New features
 ------------
 
+* DatasetAlgebra: add and subtract data of second dataset to/from dataset
+
 * ExtractCommonRange for 1D and 2D datasets
 
 * SinglePlotter2D: